foi criado o sistema de deletar e atualizar publicação

+++ routes/web.php
@@ -19,7 +19,9 @@ Route::get('/', [QuilombController::class, 'index']);
 Route::get('/users/publiction', [QuilombController::class, 'publiction'])->middleware('auth');
 Route::get('/users/{id}', [QuilombController::class, 'show']);
 Route::post('/users', [QuilombController::class, 'store']);
-Route::delete('/users/{id}', [QuilombController::class, 'destroy']);
+Route::delete('/users/{id}', [QuilombController::class, 'destroy'])->middleware('auth');
+Route::get('/users/edit/{id}', [QuilombController::class, 'edit'])->middleware('auth');
+Route::put('/users/update/{id}', [QuilombController::class, 'update'])->middleware('auth');
 
 Route::get('/dashboard', [QuilombController::class, 'dashboard'])->middleware('auth');
